Delete old test class
